doc: requirements: reword and readd MPFR URL
In 3401f947d86 (Improve documentation of GDB build requirements and
options, 2023-04-05), the "Requirements" section of the doc was reworked
to reduce repetitions and move MPFR from "Tools/Packages Optional for
Building GDB" to "Tools/Packages Necessary for Building GDB".

In the process, the URL of the MPFR homepage was forgotten. Add it
back.

Also reword slightly the first paragraph of the new  "Libraries"
subsection under "Tools/Packages Necessary for Building GDB".